Split code into multiple files for each screen
