+++ budgie.go
@@ -49,21 +49,6 @@ func main() {
 
 }
 
-type homeScreenModel struct {
-	choices  []string         // items on list
-	cursor   int              // which item our cursor is pointing at
-	selected map[int]struct{} // which items are selected
-}
-
-type insertCSVScreenModel struct {
-	filename string
-}
-
-type insertingCSVScreenModel struct {
-	filename string
-	reader   *csv.Reader
-}
-
 const (
 	insertCsvData = iota
 	deleteCsvData = iota
@@ -71,190 +56,6 @@ const (
 	deleteEntry   = iota
 	findEntry     = iota
 )
-
-// ------------------------------------------------------------
-func createInsertingCSVScreenmodel(filename string, reader *csv.Reader) insertingCSVScreenModel {
-	return insertingCSVScreenModel{
-		filename: filename,
-		reader:   reader,
-	}
-}
-
-func (m insertingCSVScreenModel) Init() tea.Cmd {
-	return nil
-}
-
-func (m insertingCSVScreenModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-
-	switch msg := msg.(type) {
-
-	case tea.KeyMsg:
-
-		switch msg.String() {
-
-		case "ctrl+c":
-			return createHomeScreenModel(), nil
-		}
-	}
-
-	return m, nil
-}
-
-func (m insertingCSVScreenModel) View() string {
-	s := "Press Ctrl + C to go back to home screen\n\n"
-	s += processCSV(m.reader)
-	return s
-}
-
-// ------------------------------------------------------------
-func createInsertCSVScreenModel() insertCSVScreenModel {
-	return insertCSVScreenModel{
-		filename: "",
-	}
-}
-
-func (m insertCSVScreenModel) Init() tea.Cmd {
-	return nil
-}
-
-func (m insertCSVScreenModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-
-	case tea.KeyMsg:
-
-		switch msg.String() {
-
-		case "ctrl+c":
-			return createHomeScreenModel(), nil
-
-		case "backspace":
-			sz := len(m.filename)
-			if sz >= 1 {
-				m.filename = m.filename[:sz-1]
-			}
-
-		case "enter":
-			data, err := readCSV(m.filename)
-			if err != nil {
-				fmt.Println("Error reading file! ", err)
-				return m, tea.Quit
-			}
-			reader, err := createCSVReader(data)
-			if err != nil {
-				fmt.Println("Error creating CSV reader: ", err)
-				return m, tea.Quit
-			}
-			m := createInsertingCSVScreenmodel(m.filename, reader)
-			return m, nil
-
-		default:
-			m.filename += msg.String()
-
-		}
-	}
-
-	return m, nil
-}
-
-func (m insertCSVScreenModel) View() string {
-
-	s := "Press Ctrl + C to go back to home screen\n\n Filename: "
-	s += style.Render(m.filename)
-
-	return s
-}
-
-// -----------------------------------------------------------
-func createHomeScreenModel() homeScreenModel {
-	return homeScreenModel{
-		choices:  []string{"Insert csv data", "Delete csv data", "Insert entry", "Delete entry", "Find entry"},
-		selected: make(map[int]struct{}), // map of int to struct
-	}
-}
-
-func (m homeScreenModel) Init() tea.Cmd {
-	return nil
-}
-
-func (m homeScreenModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-
-	// Is it a key press?
-	case tea.KeyMsg:
-
-		// Cool, what was the actual key pressed?
-		switch msg.String() {
-
-		// These keys should exit the program.
-		case "ctrl+c", "q":
-			return m, tea.Quit
-
-		// The "up" and "k" keys move the cursor up
-		case "up", "k":
-			if m.cursor > 0 {
-				m.cursor--
-			}
-
-		// The "down" and "j" keys move the cursor down
-		case "down", "j":
-			if m.cursor < len(m.choices)-1 {
-				m.cursor++
-			}
-
-		// The "enter" key and the spacebar (a literal space) toggle
-		// the selected state for the item that the cursor is pointing at.
-		case "enter", " ":
-
-			switch m.cursor {
-			case insertCsvData:
-				return createInsertCSVScreenModel(), nil
-			case deleteCsvData:
-			case insertEntry:
-			case deleteEntry:
-			case findEntry:
-			}
-
-			_, ok := m.selected[m.cursor]
-			if ok {
-				delete(m.selected, m.cursor)
-			} else {
-				m.selected[m.cursor] = struct{}{}
-				fmt.Println(m.selected)
-			}
-		}
-	}
-
-	// Return the updated model to the Bubble Tea runtime for processing.
-	// Note that we're not returning a command.
-	return m, nil
-}
-
-func (m homeScreenModel) View() string {
-	// The header
-	s := "What would you like to do?\n\n"
-
-	content := ""
-	// Iterate over our choices
-	for i, choice := range m.choices {
-
-		// Is the cursor pointing at this choice?
-		cursor := " " // no cursor
-		if m.cursor == i {
-			cursor = ">" // cursor!
-		}
-
-		// Render the row
-		content += fmt.Sprintf("%s %s\n", cursor, choice)
-	}
-
-	s += style.Render(content)
-
-	// The footer
-	s += "\nPress q to quit.\n"
-
-	// Send the UI for rendering
-	return s
-}
 
 func readCSV(filename string) ([]byte, error) {
 	f, err := os.Open(filename)
